extracted the code to composables
